TechTalk refactoring fixes
