+++ app/src/main/res/layout/main_activity.xml
@@ -16,10 +16,6 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent">
 
-<!--        <include-->
-<!--            android:id="@+id/cardViewContent"-->
-<!--            layout="@layout/incoming_order" />-->
-
     </ru.commandos.diner.delivery.view.IncomingOrderCardView>
 
     <TextView
